% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xlsxboost.R
\name{xlsx.addHeader}
\alias{xlsx.addHeader}
\title{Adds a header to a .xlsx file}
\usage{
xlsx.addHeader(wb, sheet, value = "Header", level = 1, color = "#FFFFFF",
  startRow = NULL, startCol = 2)
}
\arguments{
\item{wb}{The excel workbook}

\item{sheet}{The excel sheet (created by xlsxBoost)}

\item{value}{The header value specified}

\item{level}{The HTML header level}

\item{color}{The hex color code}

\item{startRow}{The row to start the header on}

\item{startCol}{The column to start the header on}
}
\description{
Makes use of numerous parameters including the sheet, text, HTML heading level, and color
to display a header on an excel workbook
}
\examples{
wb <- createWorkbook(type="xlsx")
sheet <- xlsx::createSheet(wb, sheetName = "example")
xlsx.addHeader(wb, sheet, value="Header 1",level=1, color="black")
saveWorkbook(wb, "examples_add_header.xlsx")
xlsx.openFile("examples_add_header.xlsx") # view the file
}

