% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xlsxboost.R
\name{xlsx.addPlot}
\alias{xlsx.addPlot}
\title{Adds a plot to a .xlsx file}
\usage{
xlsx.addPlot(wb, sheet, plotFunction, startRow = NULL, startCol = 2,
  width = 480, height = 480, ...)
}
\arguments{
\item{wb}{The excel workbook}

\item{sheet}{The excel sheet (created by xlsxboost)}

\item{plotFunction}{The function (i.e. ggplot2(...)) to show in Excel}

\item{startRow}{The row to start the header on}

\item{startCol}{The column to start the header on}

\item{width}{The width of the plot}

\item{height}{The height of the plot}
}
\description{
Makes use of numerous parameters including the sheet, text, start row/column, and
width/height in order to display a plot in an excel workbook provided a plot function.
}
\examples{
wb <- createWorkbook(type="xlsx")
sheet <- xlsx::createSheet(wb, sheetName = "example")
x = seq(from = -10*pi, to = 10*pi, by = 0.1)
pf <- function() {plot(sin(x))}
xlsx.addPlot(wb, sheet, pf)
saveWorkbook(wb, "example2.xlsx")
xlsx.openFile("example2.xlsx") # view the file
}

