% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xlsxboost.R
\name{xlsx.initialize}
\alias{xlsx.initialize}
\title{Package initialization function}
\usage{
xlsx.initialize(initializeRJava = FALSE)
}
\arguments{
\item{initializeRJava}{Set true if you are experiencing problems with initialization
of RJava (win32/64); attempts will be made to rectify installation. Default setting is false.}
}
\description{
Initializes package by preventing usage of awt on mac; to be called before use
of any functions herein.
}
\examples{
xlsx.initialize() # we may now use functionality exposed in package 'xlsxboost'
}

