% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xlsxboost.R
\name{xlsx.openFile}
\alias{xlsx.openFile}
\title{Opens a .xlsx file provided a relative filepath from getwd()}
\usage{
xlsx.openFile(filename = NULL)
}
\arguments{
\item{filename}{the name of the file to open}
}
\description{
Platform-independent method to open a .xlsx file; requires a relative filepath from getwd()
as well as capability of opening the document from terminal (Excel / LibreOffice / etc.)
}
\examples{
require('xlsx')
wb <- createWorkbook(type="xlsx")
sheet <- xlsx::createSheet(wb, sheetName = "example")
saveWorkbook(wb, "test.xlsx")
xlsx.openFile("test.xlsx")
}

