\name{xlsxboost-package}
\alias{xlsxboost-package}
\alias{xlsxboost}
\docType{package}
\title{
Provides useful extensions for dealing with excel files in R
}
\description{
Provides useful utilities including creating sheets, adding headers,
opening files, and more to work with .xlsx and .xls files in R
}
\details{
\tabular{ll}{
Package: \tab xlsxboost\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-08\cr
License: \tab MIT \cr
}
Important functions including adding header files and locally opening .xlsx files
}
\author{
Manan Shah

Maintainer: Manan Shah <manan.shah.777@gmail.com>
}
\keyword{xlsx, data}