% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xlsxboost.R
\name{xlsx.initializeRJava}
\alias{xlsx.initializeRJava}
\title{RJava initialization function}
\usage{
xlsx.initializeRJava()
}
\description{
Initializes RJava on a system by setting JAVA_HOME = ""
}
\examples{
xlsx.initializeRJava() # initializes JAVA_HOME to "" for proper installation
}

